v0.3
Enlarge cards when clicked for 0.5s, shrink back when dragged.

+++ scripts/main/card_base.gd
@@ -4,14 +4,10 @@ var card
 var pos
 var _z = 0
 var type = "hand"
+var timer = Timer.new()
 
 
 func zoom():
-	if (type=="dead"):
-		if (get_node("Animation").get_current_animation()!="fade_out"):
-			get_node("Animation").play("fade_out")
-		return
-	
 	get_node("TweenScale").remove_all()
 	get_node("TweenScale").interpolate_property(self,"scale",get_scale(),Vector2(0.3*sign(get_scale().x),0.3),0.25,Tween.TRANS_LINEAR,Tween.EASE_IN_OUT)
 	if (pos!=null):
@@ -21,24 +17,27 @@ func zoom():
 	get_node("SoundTap").play()
 
 func unzoom():
-	if (type=="dead"):
-		if (get_node("Animation").get_current_animation()!="fade_out"):
-			get_node("Animation").play("fade_out")
-	get_node("TweenScale").remove_all()
-	get_node("TweenScale").interpolate_property(self,"scale",get_scale(),Vector2(0.15*sign(get_scale().x),0.15),0.25,Tween.TRANS_LINEAR,Tween.EASE_IN_OUT)
-	if (pos!=null):
-		get_node("TweenScale").interpolate_property(self,"global_position",get_global_position(),pos,0.25,Tween.TRANS_LINEAR,Tween.EASE_IN_OUT)
-	get_node("TweenScale").start()
+	timer.stop()
+	if (get_scale().x>0.151 || get_scale().y>0.151):
+		get_node("TweenScale").remove_all()
+		get_node("SoundUntap").play()
+		get_node("TweenScale").interpolate_property(self,"scale",get_scale(),Vector2(0.15*sign(get_scale().x),0.15),0.25,Tween.TRANS_LINEAR,Tween.EASE_IN_OUT)
+		if (pos!=null):
+			get_node("TweenScale").interpolate_property(self,"global_position",get_global_position(),pos,0.25,Tween.TRANS_LINEAR,Tween.EASE_IN_OUT)
+		get_node("TweenScale").start()
+	else:
+		set_scale(Vector2(0.15*sign(get_scale().x),0.15))
 	set_z_index(_z)
-	get_node("SoundUntap").play()
 
 func _gui_input(event):
 	if (type=="dead"):
 		return
 	
-	if (event.is_action_pressed("RMB")):
+	if (event.is_action_pressed("LMB")):
+		timer.start()
+	elif (event.is_action_pressed("RMB")):
 		zoom()
-	elif (event.is_action_released("RMB")):
+	elif (event.is_action_released("LMB") || event.is_action_released("RMB")):
 		unzoom()
 
 func selected():
@@ -53,4 +52,8 @@ func _ready():
 	get_node("Button").connect("pressed",self,"selected")
 	get_node("Button").connect("gui_input",self,"_gui_input")
 #	get_node("Button").connect("mouse_entered",self,"zoom")
-#	get_node("Button").connect("mouse_exited",self,"unzoom")
+	get_node("Button").connect("mouse_exited",self,"unzoom")
+	timer.set_wait_time(0.5)
+	timer.set_one_shot(true)
+	add_child(timer)
+	timer.connect("timeout",self,"zoom")
